Fix `SensorListener`'s sampling rate bug
This commit will fix the bug in `SensorListener` that caused bad sampling on some devices that
completely disregard the sensor's `samplingPeriodUs`.

Some real-life devices completely disregard `SensorEventListener.registerListener`'s parameter
`samplingPeriodUs` for a fixed sampling rate that may not match the desired one. For example, if
we want to sample at a rate of 50Hz, but the device sends us data at 100Hz the sample window
will be filled in half the time messing with the whole sampling timing.

This commit redesigns `SensorListener` to discard events faster than the desired sampling rate,
it's not the best looking solution but it seems the better now.