Sensorlib: Add `Mocks` helper file
This commit will add an helper file inside sensorslib's unit tests containing some
usefull mocks used by the other tests.

For the moment this file is local the the sensorslib module, but in the future it
will be extracted in a common module and used all the other modules of the project.