Sensorlib: Fix bug in `SensorWindowBuffer` producing bad windows
This commit fixes a bug in `SensorWindowBuffer` causing it to produce a window
when in reality it was only half-done.

This bug was caused by a check in `SensorWindowBuffer`'s methods `pushAccelerometer`
and `pushGyroscope` that was checking only one of the buffer indeces for overflow and
not both of them.

If we where receiving all the gyroscope samples, and only part of the accelerometer
ones (or vice-versa) this was firing not waiting for all the accelerometer samples
and resulting in a corrupted window with some zeros in it.

This commit olso adds a parameter to `SensorSampleFilter` to mark the name of the
sensor being filtered in debug logs.