Gradle: Move all dependencies to versionCatalog
This commit changes the way the dependencies for all modules are declared.
Introduce the use of `versionCatalog` inside `settings.gradle` to store all
the versions and libraries used throughout the project's sub-modules.

In the next commit we will switch to using a `libs.versions.toml` file to
manage those versions even better.